ola ke ase

+++ application/controllers/Reproductor.php
@@ -85,7 +85,7 @@ class Reproductor extends CI_Controller {
     public function UploadFromURL() {
         $data = $this->input->post('BibliotecaMusicalStream');
         $cancion = strip_tags($this->carpeta . $data['Titulo'] . '_-_' . $data['Artista'] . '.mp3');
-        if (!file_exists($cancion)) {
+        if (!file_exists($cancion) && !$this->Base->getDataWhere(Reproductor::$TBLBibliotecaMusical, array('Origen' => $data['URL']))) {
             $url = exec('casperjs C:\SITES\CasperJS\offliberty.js --url="' . $data['URL'] . '"');
             file_put_contents($cancion, fopen($url, 'r'));
             $this->EditID3($cancion, $data, true);
